chore: create useActiveAnchors hook and cleanup
